+++ src/slogo/model/Commands/TurtleCommands/SetHeading.java
@@ -4,7 +4,7 @@ import slogo.model.TreeNode;
 
 public class SetHeading extends TurtleCommand {
 
-  private int degrees;
+  private double degrees;
 
   public SetHeading(String name) {
     super(name);
@@ -12,8 +12,8 @@ public class SetHeading extends TurtleCommand {
 
   @Override
   public void doCommand(TreeNode commandNode) {
-    degrees = (int)Double.parseDouble(getParamList().get(0));
-    int returnDegrees = degrees - turtle.getDegree(); //value that needs to be sent back
+    degrees = Double.parseDouble(getParamList().get(0));
+    double returnDegrees = degrees - turtle.getDegree(); //value that needs to be sent back
     turtle.setDegree(degrees % 360);
     commandNode.setData("" + returnDegrees);
   }
